Minor code cleanup of things seen while reviewing changes, and convert a few configs to longs that got missed

+++ src/generators/java/mekanism/generators/common/content/fission/FissionReactorMultiblockData.java
@@ -136,7 +136,7 @@ public class FissionReactorMultiblockData extends MultiblockData implements IVal
     @ContainerSync
     private boolean forceDisable;
 
-    private int cooledCoolantCapacity;
+    private long cooledCoolantCapacity;
     private long heatedCoolantCapacity;
     private long fuelCapacity;
